+++ mingling/src/res.rs
@@ -1,0 +1,2 @@
+mod exit_code;
+pub use exit_code::*;