+++ mingling/src/res/exit_code.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Default, Clone, Copy)]
+pub struct ExitCode {
+    pub exit_code: i32,
+}