+++ mingling/src/setups.rs
@@ -1,0 +1,11 @@
+mod basic;
+pub use basic::*;
+
+mod exit_code;
+pub use exit_code::*;
+
+#[cfg(feature = "general_renderer")]
+mod general_renderer;
+
+#[cfg(feature = "general_renderer")]
+pub use general_renderer::*;